modifications to allow for bigger INT_TYPE then the standard 32 bit or 'signed long': Gmp.mpz modified to create objects from bigger native int types (get_mpz_new) sprintf and configure modified so it can print bigger ints with then '%d', tests for '%llx' among others main configure and global.h to allow type setting with --with-long-long-int et al (default to not use long long is still in global.h) various warnings fixed (not all)

+++ src/modules/Image/colors.c
@@ -653,8 +653,8 @@ static void image_color_hex(INT32 args)
 
       sh=4*(sizeof(COLORTYPE)*2-i);
       if (sh>0)
-	 sprintf(buf,"#%0*x%0*x%0*x",i,THIS->rgb.r>>sh,
-		 i,THIS->rgb.g>>sh,i,THIS->rgb.b>>sh); 
+	 sprintf(buf,"#%0*x%0*x%0*x",i,(unsigned)(THIS->rgb.r>>sh),
+		 i,(unsigned)(THIS->rgb.g>>sh),i,(unsigned)(THIS->rgb.b>>sh)); 
       else
       {
 	 unsigned INT32 r=THIS->rgbl.r;
@@ -669,7 +669,7 @@ static void image_color_hex(INT32 args)
 	    sh=0;
 	 }
 	 sprintf(buf,"#%0*x%0*x%0*x",
-		 i,r>>sh,i,g>>sh,i,b>>sh);
+		 i,(unsigned)(r>>sh),i,(unsigned)(g>>sh),i,(unsigned)(b>>sh));
       }
    }
    else
